Leave standard margin on QR code
Removing the margin from the QR code image is okay when you're displaying the
image on a white background and adding margin another way, e.g. using CSS as in
the web client. However, if you need to use the save the generated image for use
in another medium, especially one with a dark background, the lack of whitespace
around the grid can slow detection or make it unreadable, depending on the
algorithm used and quality of the image taken by the camera. It's safest to
leave the default margin in.